refactor(telegram): extract jsonl_kernel + rename jsonl_watcher to claude_watcher (#258)
Pure refactor for the per-backend session-reader triad (claude/codex/gemini)
landing in subsequent commits. Splits the shared scaffold from Claude's
specific bits:

- `telegram/jsonl_watcher.rs` -> `telegram/claude_watcher.rs` (git mv preserves blame).
- New `telegram/jsonl_kernel.rs` hosts `find_latest_jsonl`, `read_new_lines`,
  and the polling/rotation constants (`POLL_INTERVAL_MS`, `ROTATION_STALE_SECS`).
- `claude_watcher.rs` keeps its Claude-specific `extract_assistant_text` and
  the watch_loop scaffold; reads the kernel primitives via `crate::telegram::jsonl_kernel`.
- `bridge::spawn_bridge` calls `super::claude_watcher::spawn_watch_task` instead of
  the renamed module.

No behavior change: the H2 truncation-skip semantics lands in commit 5
(preamble scan + truncation-skip fix). Today's `read_new_lines` keeps the
existing reset-to-0 behavior to keep this commit a pure refactor.
