+++ src-tauri/src/telegram/mod.rs
@@ -1,5 +1,6 @@
 pub mod api;
 pub mod bridge;
-pub mod jsonl_watcher;
+pub mod claude_watcher;
+pub mod jsonl_kernel;
 pub mod manager;
 pub mod types;
